requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
PyMuPDF
httpx
google-generativeai
asyncio
